fixed ainstr/cinstr codegen and allowed comments after instructions
variable addr was not being allocated because the pointer was changing
from underneath, so the unordered_map key was changed to std::string, this
also fixed the ainstr label generating the wrong instruction.

Fixed cinstr Destination::D wrongly being marked as Destination::None.

Changed the lexing of comments to allow trailing newline so that comments
can be added after an instruction.